web/requirements.txt
django==5.0.4
djangorestframework==3.15.1
drf-spectacular==0.27.2
django-debug-toolbar==4.3.0
django-environ==0.11.2
requests==2.31.0
djoser==2.2.2